fix(webpack): typescript imports in non-ts projects
