+++ packages/webpack5/src/helpers/index.ts
@@ -26,7 +26,7 @@ import {
 	getPlatform,
 	getPlatformName,
 } from './platform';
-import { readTsConfig } from './tsconfig';
+import { readTsConfig } from './typescript';
 
 // intentionally populated manually
 // as this generates nicer typings
